リファクタ、不要なインターフェイスや構造体を削除
不要なインターフェイスを削除中

インターフェイス削除中

同上

不要な箇所を削除

不要な構造体を削除

リファクタ中

リファクタ中
