+++ Editor/RoadNetwork/EditingSystem/EditingIntersection.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 89637c64e4d043ed8c56d3074c60d6e6
+timeCreated: 1735291200
