--- Editor/RoadNetwork/EditingSystem/IRoadNetworkEditingSystem.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 482859cd48d94b6c88afc7aa38e5b35f
-timeCreated: 1733822397
