+++ Editor/RoadNetwork/EditingSystem/RoadNetworkEditSceneViewGui.cs
@@ -21,14 +21,14 @@ namespace PLATEAU.Editor.RoadNetwork.EditingSystem
         public List<EditorData<RnRoadGroup>> Connections { get => roadGroupEditorData; }
 
         public RoadNetworkEditSceneViewGui(GameObject root, RnModel rnModel,
-            IRoadNetworkEditingSystem system)
+            EditingSystem system)
         {
             ReConstruct(root, rnModel, system);
         }
 
         private GameObject roadNetworkEditingSystemObjRoot;
         private RnModel roadNetwork;
-        private IRoadNetworkEditingSystem system;
+        private EditingSystem system;
         
         // 詳細編集モードかどうか
         public bool isEditingDetailMode = false;
@@ -57,7 +57,7 @@ namespace PLATEAU.Editor.RoadNetwork.EditingSystem
         /// 計算や処理に必要な要素を初期化する
         /// </summary>
         public void ReConstruct(GameObject root, RnModel rnModel,
-            IRoadNetworkEditingSystem system)
+            EditingSystem system)
         {
             Assert.IsNotNull(root);
             Assert.IsNotNull(rnModel);
@@ -237,14 +237,24 @@ namespace PLATEAU.Editor.RoadNetwork.EditingSystem
             intersectionEditorData.Clear();
         }
 
-        
-
+        /// <summary>
+        /// 描画します
+        /// </summary>
         private void Update(SceneView sceneView)
         {
             if (roadNetworkEditingSystemObjRoot == null)
                 return;
             
+            UpdateRoad(sceneView);
+            UpdateIntersection();
 
+
+            // 仮で呼び出し　描画の更新がワンテンポ遅れるため　
+            EditorUtility.SetDirty(roadNetworkEditingSystemObjRoot);
+        }
+
+        private void UpdateRoad(SceneView sceneView)
+        {
             // マウス位置に近いwayを算出
 
             if (this.roadGroupEditorData.TryGetCache<RoadGroupEditorData>("linkGroup", out var eConn) == false)
@@ -296,10 +306,85 @@ namespace PLATEAU.Editor.RoadNetwork.EditingSystem
                 //    intersectionsPoss.Add(item.RefGameObject.transform.position);
                 //}
             }
+        }
+
+        private void UpdateIntersection()
+        {
+            var editingIntersectionMod = system.EditSceneViewGui.EditingIntersectionMod;
+            if (editingIntersectionMod.Intersection == null) return;
+            
+            RoadEditSceneGUIState state = new RoadEditSceneGUIState();
+
+            var currentCamera = SceneView.currentDrawingSceneView.camera;
+            state.currentCamera = currentCamera;
 
 
-            // 仮で呼び出し　描画の更新がワンテンポ遅れるため　
-            EditorUtility.SetDirty(roadNetworkEditingSystemObjRoot);
+            
+
+            var buttonSize = 2.0f;
+
+            bool isSelectdEntablePoint = editingIntersectionMod.IsSelectdEntablePoint;
+            if (isSelectdEntablePoint == false)
+            {
+                foreach (var item in editingIntersectionMod.EnterablePoints)
+                {
+                    // 流入点の位置にボタンを表示する
+                    if (Handles.Button(item.CalcCenter(), Quaternion.identity, buttonSize, buttonSize,
+                            RoadNetworkEntarablePointButtonHandleCap))
+                    {
+                        editingIntersectionMod.SetEntablePoint(item);
+                        // 流入点が選択された
+                        break;
+                    }
+                }
+            }
+            else
+            {
+                foreach (var item in editingIntersectionMod.ExitablePoints)
+                {
+                    // 流出点の位置にボタンを表示する
+                    if (Handles.Button(item.CalcCenter(), Quaternion.identity, buttonSize, buttonSize,
+                            RoadNetworkExitablePointButtonHandleCap))
+                    {
+                        // 流出点が選択された
+                        editingIntersectionMod.SetExitablePoint(item);
+                        break;
+                    }
+                }
+            }
+
+            // Trackの生成、削除に必要な設定が済んで更新できるか？
+            if (editingIntersectionMod.CanTryUpdateTrack)
+            {
+                editingIntersectionMod.UpdateTrack();
+            }
+
+
+            // 遅延実行 コレクションの要素数などを変化させる
+            if (state.delayCommand != null)
+                state.delayCommand.Invoke();
+
+            // 変更を通知する
+            if (state.isDirtyTarget)
+            {
+                system.NotifyChangedRoadNetworkObject2Editor();
+            }
+        }
+        
+        private static void RoadNetworkExitablePointButtonHandleCap(int controlID, Vector3 position,
+            Quaternion rotation, float size, EventType eventType)
+        {
+            if (eventType == EventType.Repaint)
+                Handles.color = Color.blue;
+            Handles.SphereHandleCap(controlID, position, rotation, size, eventType);
+        }
+        
+        private static void RoadNetworkEntarablePointButtonHandleCap(int controlID, Vector3 position,
+            Quaternion rotation, float size, EventType eventType)
+        {
+            if (eventType == EventType.Repaint)
+                Handles.color = Color.red;
+            Handles.SphereHandleCap(controlID, position, rotation, size, eventType);
         }
 
         private RoadNetworkEditorGizmos GetRoadNetworkEditorGizmos()
@@ -308,7 +393,7 @@ namespace PLATEAU.Editor.RoadNetwork.EditingSystem
             return roadNetworkEditingSystemObjRoot.GetComponent<RoadNetworkEditorGizmos>();
         }
 
-        public void Setup(EditorData<RnIntersection> data)
+        public void SetupIntersection(EditorData<RnIntersection> data)
         {
             editingIntersection.SetTarget(data);
             editingIntersection.Activate(true);
@@ -332,123 +417,6 @@ namespace PLATEAU.Editor.RoadNetwork.EditingSystem
             new RoadReproducer().Generate(new RrTargetRoadBases(roadNetwork, roads.ToArray()), CrosswalkFrequency.All);
         }
 
-        public class EditingIntersection
-        {
-            public bool SetTarget(EditorData<RnIntersection> intersection)
-            {
-                if (this.intersection == intersection)
-                    return false;
-
-                if (intersection == null)
-                {
-                    Activate(false);
-                    return true;
-                }
-
-                this.intersection = intersection;
-                return true;
-            }
-
-            public void Activate(bool activate)
-            {
-                this.activate = activate;
-            }
-
-            public IReadOnlyCollection<RnNeighbor> EnterablePoints
-            {
-                get
-                {
-                    var d = intersection.ReqSubData<EnterablePointEditorData>();
-                    return d.Points;
-                }
-            }
-
-            public IReadOnlyCollection<RnNeighbor> ExitablePoints
-            {
-                get
-                {
-                    var d = intersection.ReqSubData<ExitablePointEditorData>();
-                    return d.Points;
-                }
-            }
-
-            /// <summary>
-            /// 流入点と流出点を返す
-            /// </summary>
-            public (RnNeighbor, RnNeighbor) SelectedPoints
-            {
-                get => (selectEntablePoint, selectExitablePoint);
-            }
-
-            public bool IsSelectdEntablePoint
-            {
-                get => selectEntablePoint != null;
-            }
-
-            public bool CanTryUpdateTrack
-            {
-                get => selectEntablePoint != null && selectExitablePoint != null;
-            }
-
-            public void SetEntablePoint(RnNeighbor neighbor)
-            {
-                Assert.IsNotNull(neighbor);
-
-                // 選択中の交差点に含まれているか
-                Assert.IsTrue(EnterablePoints.Contains(neighbor));
-
-                selectEntablePoint = neighbor;
-            }
-
-            public void SetExitablePoint(RnNeighbor neighbor)
-            {
-                Assert.IsNotNull(neighbor);
-
-                // 選択中の交差点に含まれているか
-                Assert.IsTrue(ExitablePoints.Contains(neighbor));
-
-                selectExitablePoint = neighbor;
-            }
-
-            /// <summary>
-            /// 選択状態やトラックの有無で処理を分岐する
-            /// </summary>
-            public void UpdateTrack()
-            {
-                Assert.IsTrue(CanTryUpdateTrack);
-
-                var track = intersection.Ref.FindTrack(selectEntablePoint, selectExitablePoint);
-                if (track != null)
-                    intersection.Ref.RemoveTrack(track);
-                else
-                    intersection.Ref.TryAddOrUpdateTrack(selectEntablePoint, selectExitablePoint);
-
-                selectEntablePoint = null;
-                selectExitablePoint = null;
-            }
-
-            public void RemoveTarck()
-            {
-                selectEntablePoint = null;
-                selectExitablePoint = null;
-            }
-
-            //public void CreateSubData()
-            //{
-            //    intersection.ClearSubData();
-
-            //    var enterablePointEditorData = EnterablePointEditorData.Create(intersection); 
-            //    intersection.TryAdd(enterablePointEditorData);
-            //}
-
-
-            private EditorData<RnIntersection> intersection;
-            private bool activate = false;
-
-            private RnNeighbor selectEntablePoint = null;
-            private RnNeighbor selectExitablePoint = null;
-
-            // private bool isShapeEditingMode = false;
-        }
+        
     }
 }
